+++ packages/extension-integrationtest/jest.config.js
@@ -1,0 +1,14 @@
+function isCI() {
+  return !!process.env['CI'];
+}
+
+const baseTimeout = 20_000; // 20s
+
+/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
+module.exports = {
+  runner: 'vscode',
+  modulePathIgnorePatterns: ['.vscode-test/'],
+  testMatch: ['**/out/**/*.test.js'],
+  testTimeout: isCI() ? baseTimeout * 10 : baseTimeout,
+  verbose: true,
+};