+++ packages/extension/src/integrationTest/testCommands.ts
@@ -1,0 +1,14 @@
+export const enum TestCommands {
+  GetDecorationSetterRecording = 'rxjs-debugging-for-vs-code.command.test.getDecorationSetterRecording',
+}
+
+export interface ITestCommandTypes {
+  [TestCommands.GetDecorationSetterRecording]: (
+    file: string,
+    decorationType: 'liveLog' | 'logPoints'
+  ) => ReadonlyArray<{
+    decorationType: string;
+    ranges: ReadonlyArray<string>;
+    options: ReadonlyArray<string>;
+  }>;
+}